# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gestor de Estoque API",
    "description": "API REST para gestão de empresas, usuários, produtos, combos, promoções e movimentações de estoque.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Ambiente local" }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Autenticar usuário e obter token JWT",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Autenticado com sucesso",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/empresas": {
      "get": {
        "tags": ["Empresas"],
        "summary": "Buscar a empresa do usuário autenticado",
        "description": "Retorna, em uma lista de um item, a empresa do usuário autenticado (não há listagem entre empresas diferentes).",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista contendo a empresa do usuário autenticado",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Empresa" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Empresas"],
        "summary": "Criar empresa (cadastro público, self-service)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmpresaCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Empresa criada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Empresa" } } }
          },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/empresas/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Empresas"],
        "summary": "Buscar empresa por ID (somente a própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Empresa encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Empresa" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Empresas"],
        "summary": "Atualizar empresa (somente a própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmpresaUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Empresa atualizada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Empresa" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Empresas"],
        "summary": "Remover empresa (soft delete, somente a própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/catalogo/produtos": {
      "get": {
        "tags": ["Catálogo"],
        "summary": "Listar produtos do catálogo público de uma empresa",
        "description": "Rota pública. Retorna apenas produtos ativos e visíveis no catálogo, com campos seguros para exibição ao cliente final (sem precoAtacado, estoque ou dados internos).",
        "parameters": [
          { "name": "empresaId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "Informe empresaId ou slug." },
          { "name": "slug", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Informe empresaId ou slug." }
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos do catálogo",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProdutoCatalogo" } } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/catalogo/combos": {
      "get": {
        "tags": ["Catálogo"],
        "summary": "Listar combos do catálogo público de uma empresa",
        "description": "Rota pública. Retorna apenas combos ativos e visíveis no catálogo.",
        "parameters": [
          { "name": "empresaId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "Informe empresaId ou slug." },
          { "name": "slug", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Informe empresaId ou slug." }
        ],
        "responses": {
          "200": {
            "description": "Lista de combos do catálogo",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ComboCatalogo" } } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/combos": {
      "get": {
        "tags": ["Combos"],
        "summary": "Listar combos da própria empresa",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de combos",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Combo" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Combos"],
        "summary": "Criar combo",
        "description": "O empresaId é sempre o da empresa do usuário autenticado; qualquer valor enviado no corpo é ignorado.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComboCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Combo criado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Combo" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/combos/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Combos"],
        "summary": "Buscar combo por ID (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Combo encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Combo" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Combos"],
        "summary": "Atualizar combo (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComboUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Combo atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Combo" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Combos"],
        "summary": "Remover combo (soft delete, somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/movimentacoes/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Movimentações"],
        "summary": "Buscar movimentação por ID (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Movimentação encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MovimentacaoEstoque" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Movimentações"],
        "summary": "Remover movimentação (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token obtido em POST /api/auth/login. Envie como 'Authorization: Bearer <token>'."
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "Message": {
        "description": "Mensagem de confirmação",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
      },
      "Error": {
        "description": "Erro",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Message": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "Empresa": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "slug": { "type": "string" },
          "logo": { "type": "string", "nullable": true },
          "banner": { "type": "string", "nullable": true },
          "descricao": { "type": "string", "nullable": true },
          "telefone": { "type": "string", "nullable": true },
          "instagram": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "EmpresaCreate": {
        "type": "object",
        "required": ["nome", "slug"],
        "properties": {
          "nome": { "type": "string" },
          "slug": { "type": "string" },
          "logo": { "type": "string" },
          "banner": { "type": "string" },
          "descricao": { "type": "string" },
          "telefone": { "type": "string" },
          "instagram": { "type": "string" }
        }
      },
      "EmpresaUpdate": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "slug": { "type": "string" },
          "logo": { "type": "string" },
          "banner": { "type": "string" },
          "descricao": { "type": "string" },
          "telefone": { "type": "string" },
          "instagram": { "type": "string" }
        }
      },
...
      "ProdutoCatalogo": {
        "type": "object",
        "description": "Versão pública e filtrada do produto, sem precoAtacado, estoque ou dados internos.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "codigo": { "type": "string" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "precoVarejo": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "imagens": { "type": "array", "items": { "$ref": "#/components/schemas/ProdutoImagem" } }
        }
      },
      "ComboCatalogo": {
        "type": "object",
        "description": "Versão pública e filtrada do combo.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "preco": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "itens": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "quantidade": { "type": "integer" },
                "produto": { "$ref": "#/components/schemas/ProdutoCatalogo" }
              }
            }
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "senha"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "senha": { "type": "string", "format": "password" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "JWT válido por 7 dias. Envie em requisições privadas como 'Authorization: Bearer <token>'." },
          "usuario": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "nome": { "type": "string" },
              "email": { "type": "string", "format": "email" },
              "role": { "type": "string", "enum": ["ADMIN"] },
              "empresaId": { "type": "string", "format": "uuid" }
            }
          }
        }
      },
...
      "ProdutoImagem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "url": { "type": "string" },
          "alt": { "type": "string", "nullable": true },
          "ordem": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Produto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "codigo": { "type": "string" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "precoVarejo": { "type": "number", "format": "double" },
          "precoAtacado": { "type": "number", "format": "double" },
          "estoque": { "type": "integer" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "visivelCatalogo": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "imagens": { "type": "array", "items": { "$ref": "#/components/schemas/ProdutoImagem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "ComboItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "comboId": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "quantidade": { "type": "integer" },
          "produto": { "$ref": "#/components/schemas/Produto" }
        }
      },
      "Combo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "preco": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "visivel": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "itens": { "type": "array", "items": { "$ref": "#/components/schemas/ComboItem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ComboCreate": {
        "type": "object",
        "required": ["empresaId", "nome", "preco"],
        "properties": {
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "preco": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string" },
          "ordemCatalogo": { "type": "integer", "default": 0 },
          "destaque": { "type": "boolean", "default": false },
          "ativo": { "type": "boolean", "default": true },
          "visivel": { "type": "boolean", "default": true }
        }
      },
      "ComboUpdate": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "preco": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string" },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "visivel": { "type": "boolean" }
        }
      },
...
      "UsuarioResumo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "MovimentacaoEstoque": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "usuarioId": { "type": "string", "format": "uuid" },
          "tipo": { "type": "string", "enum": ["ENTRADA", "SAIDA", "AJUSTE"] },
          "quantidade": { "type": "integer" },
          "motivo": { "type": "string", "nullable": true },
          "produto": { "$ref": "#/components/schemas/Produto" },
          "usuario": { "$ref": "#/components/schemas/UsuarioResumo" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}